+++ Cargo.toml
@@ -18,8 +18,7 @@ clap = { version = "4.5", optional = true, features = ["derive"] }
 crc32fast = "1.4.2"
 crossbeam-channel = "0.5.13"
 crossbeam-skiplist = "0.1.3"
-derive_more = { version = "1.0.0", features = ["display"] }
-dotenv = "0.15.0"
+dotenvy = { version = "0.15.7", optional = true }
 env_logger = { version = "0.11", optional = true }
 fail-parallel = "0.5.1"
 flatbuffers = "24.3.25"
@@ -30,7 +29,7 @@ parking_lot = "0.12.3"
 siphasher = "1"
 thiserror = "1.0.63"
 tokio = { version = "1.40.0", features = ["macros", "sync", "rt", "rt-multi-thread"] }
-ulid = "1.1.3"
+ulid = { version = "1.1.3", features = ["serde"] }
 rand = "0.8.5"
 rand_xorshift = "0.3.0"
 log = {  version = "0.4.22", features = ["std"] }
@@ -54,7 +53,7 @@ tempfile = "3.3"
 [features]
 default = ["aws"]
 aws = ["object_store/aws"]
-cli = ["clap", "clap/derive"]
+cli = ["clap", "clap/derive", "dotenvy"]
 db_bench = ["aws", "env_logger", "leaky-bucket", "clap", "clap/derive", "wal_disable"]
 compression = ["snappy"]
 snappy = ["dep:snap"]
